Feat: 기존 사진 삭제만 하는 기능 추가

+++ client/src/components/PostCard.tsx
@@ -1,14 +1,11 @@
-import { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 import { PostCardProps } from '../data/constants/types/post';
-import { useForm, SubmitHandler } from 'react-hook-form';
-import { editPostFormFields } from '../data/constants/types/post';
 import { useRecoilState } from 'recoil';
 import { userPostsState } from '../utils/state';
-import { patchPost, getPosts, deletePost, postPost } from '../utils/axios';
-import ImageCropper from './ImageCropper';
-import { CropperState } from '../data/constants/types/image';
+import { getPosts, deletePost } from '../utils/axios';
+import PostCardEdit from './PostCardEdit';
+import { useNavigate } from 'react-router-dom';
 
-import dayjs from 'dayjs';
 import {
 	__Card,
 	__CardImage,
@@ -46,6 +43,7 @@ import {
 	__CancelSelectButton,
 	__CropImageButton,
 	__SelectImageButton,
+	__DeleteOriginalImageButton,
 	__ImageOverlay,
 	__Form,
 	__InputContainer,
@@ -64,24 +62,16 @@ import {
 	__PostFormInfoCol,
 	__PostFormControlsCol,
 	__HeaderInputContainer,
+	__CardDate,
+	__NoImage,
 } from '../styles/__components/__PostCard';
-import { useNavigate } from 'react-router-dom';
 
 export default function PostCard({ post }: PostCardProps) {
 	const [postsState, setPostsState] = useRecoilState(userPostsState);
 	const [showModal, setShowModal] = useState(false);
-	const [selectedImage, setSelectedImage] = useState<File | null>(null);
-	const [selectedImageURL, setSelectedImageURL] = useState<string | null>(null);
-	const [croppedImage, setCroppedImage] = useState<File | null>(null);
-	const [croppedImageURL, setCroppedImageURL] = useState<string | null>(null);
-	const [cropper, setCropper] = useState<CropperState | null>(null);
-	const [showCropperModal, setShowCropperModal] = useState(false);
 	const [loading, setLoading] = useState(false);
-
+	const [isEditing, setIsEditing] = useState(false);
 	const navigate = useNavigate();
-	const closeImageCropper = () => {
-		setShowCropperModal(false);
-	};
 
 	const handleShowModal = () => {
 		setShowModal(true);
@@ -91,90 +81,14 @@ export default function PostCard({ post }: PostCardProps) {
 		setShowModal(false);
 	};
 
-	const [isEditing, setIsEditing] = useState(false);
-
-	useEffect(() => {
-		if (croppedImage) {
-			setCroppedImageURL(URL.createObjectURL(croppedImage));
-		}
-
-		return () => {
-			if (croppedImageURL) {
-				URL.revokeObjectURL(croppedImageURL);
-			}
-		};
-	}, [croppedImage]);
-
-	useEffect(() => {
-		if (selectedImage) {
-			const url = URL.createObjectURL(selectedImage);
-			setSelectedImageURL(url);
-		}
-	}, [selectedImage]);
-
-	const onFileChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-		if (event.target.files && event.target.files.length > 0) {
-			setCropper(null);
-			setSelectedImage(event.target.files[0]);
-			setShowCropperModal(true);
-		}
-	};
-
-	const clearEditForm = () => {
-		setIsEditing(false);
-		setSelectedImage(null);
-		setSelectedImageURL(null);
-		setCroppedImage(null);
-		setCroppedImageURL(null);
-		setCropper(null);
-		reset(post);
-	};
-
-	const {
-		register,
-		handleSubmit,
-		formState: { errors, isValid },
-		reset,
-	} = useForm<editPostFormFields>({
-		mode: 'onChange',
-		defaultValues: post,
-	});
-
-	const editPost: SubmitHandler<editPostFormFields> = async (values) => {
-		setLoading(true);
-
-		const formData = new FormData();
-		if (croppedImage) {
-			formData.append('file', croppedImage);
-		}
-		const postData = {
-			date: values.date,
-			title: values.title,
-			content: values.content,
-			plant: values.plant,
-		};
-		formData.append('postData', JSON.stringify(postData));
-
-		const { result, message, data } = await patchPost(post.id, formData, { width: 400, height: 500 });
-		if (result) {
-			const response = await getPosts();
-			clearEditForm();
-			setPostsState(response.data);
-			setLoading(false);
-		}
-		if (!result) {
-			setLoading(false);
-			alert(message);
-			navigate('/signin');
-		}
-	};
-
 	const handleDeletePost = async () => {
+		if (!window.confirm('게시글을 삭제 하시겠습니까?')) {
+			return;
+		}
 		setLoading(true);
 		const { result, message, data } = await deletePost(post.id);
 		if (result) {
 			const response = await getPosts();
-			clearEditForm();
 			setPostsState(response.data);
 			setLoading(false);
 			alert(message);
@@ -183,31 +97,6 @@ export default function PostCard({ post }: PostCardProps) {
 			setLoading(false);
 			alert(message);
 			navigate('/signin');
-		}
-	};
-
-	useEffect(() => {
-		clearEditForm();
-	}, [post]);
-
-	const imageInputRef = useRef<HTMLInputElement>(null);
-
-	const triggerImageSelect = () => {
-		if (imageInputRef.current) {
-			imageInputRef.current.click();
-		}
-	};
-
-	const clearImage = () => {
-		setSelectedImage(null);
-		setSelectedImageURL(null);
-		setCroppedImage(null);
-		setCroppedImageURL(null);
-	};
-
-	const resetFile = () => {
-		if (imageInputRef.current) {
-			imageInputRef.current.value = '';
 		}
 	};
 
@@ -215,7 +104,7 @@ export default function PostCard({ post }: PostCardProps) {
 		<>
 			<__CardCol>
 				<__Card onClick={handleShowModal}>
-					{post.img && <__CardImage variant='top' src={post.img} />}
+					{post.img ? <__CardImage variant='top' src={post.img} /> : <__CardDate>{post.date}</__CardDate>}
 					<__CardImageOverlay>
 						<__CardBody>
 							<__CardTitle>{post.title}</__CardTitle>
@@ -224,19 +113,18 @@ export default function PostCard({ post }: PostCardProps) {
 					</__CardImageOverlay>
 				</__Card>
 			</__CardCol>
+
 			<__Modal
 				show={showModal}
 				onHide={handleCloseModal}
 				backdrop={isEditing ? 'static' : true}
 				size={'lg'}
-				onExited={clearEditForm}
+				keyboard={!isEditing}
 				dialogClassName={'custom-modal'}>
 				{!isEditing && (
 					<__ModalHeader>
 						<__Dropdown>
-							<__DropdownToggle>
-								<__DotsIcon />
-							</__DropdownToggle>
+							<__DropdownToggle>{loading ? <__Spinner /> : <__DotsIcon />}</__DropdownToggle>
 							<__DropdownMenu>
 								<__DropdownItem
 									onClick={() => {
@@ -251,134 +139,10 @@ export default function PostCard({ post }: PostCardProps) {
 				)}
 				<__ModalBody>
 					{isEditing ? (
-						<__PostFormContainer>
-							<__Form onSubmit={handleSubmit(editPost)}>
-								<__PostFormRow>
-									<__PostFormImageCol xs={12} sm={12} md={12} lg={6}>
-										<__ImagePreviewContainer>
-											{(post.img || croppedImageURL) && (
-												<__ImagePreview src={croppedImageURL || post.img} />
-											)}
-											<__ImageOverlay>
-												<__ImageEditControlsContainer>
-													<__SelectImageButton onClick={triggerImageSelect} />
-													{selectedImage && (
-														<>
-															<__CropImageButton
-																onClick={() => {
-																	setShowCropperModal(true);
-																}}
-															/>
-															<__CancelSelectButton onClick={clearImage} />
-														</>
-													)}
-												</__ImageEditControlsContainer>
-
-												<__FileInput
-													type={'file'}
-													accept={'image/*'}
-													onChange={onFileChange}
-													onClick={resetFile}
-													ref={imageInputRef}
-												/>
-											</__ImageOverlay>
-										</__ImagePreviewContainer>
-									</__PostFormImageCol>
-
-									<__PostFormInfoCol xs={12} sm={12} md={12} lg={6}>
-										<__HeaderInputContainer>
-											<__InputContainer>
-												<__InputLabel>날짜</__InputLabel>
-												<__Input
-													type={'date'}
-													{...register('date', {
-														required: '날짜를 선택해주세요',
-														setValueAs: (value) => dayjs(value).format('YYYY-MM-DD'),
-													})}
-													min={dayjs(new Date(2020, 0, 1)).format('YYYY-MM-DD')}
-													max={dayjs(new Date()).format('YYYY-MM-DD')}
-												/>
-											</__InputContainer>
-											<__InputContainer>
-												<__InputLabel>제목</__InputLabel>
-
-												<__Input
-													type={'text'}
-													{...register('title', {
-														required: '제목을 입력해주세요.',
-														validate: {
-															emptyCheck: (value) =>
-																value.trim().length < 1
-																	? '제목을 입력해주세요.'
-																	: undefined,
-														},
-													})}
-												/>
-											</__InputContainer>
-											<__InputContainer>
-												<__InputLabel>반려식물 이름</__InputLabel>
-
-												<__Input
-													type={'text'}
-													{...register('plant', {
-														required: '반려 식물의 이름을 입력해주세요.',
-														validate: {
-															emptyCheck: (value) =>
-																value.trim().length < 1
-																	? '반려 식물의 이름을 입력해주세요.'
-																	: undefined,
-														},
-													})}
-												/>
-											</__InputContainer>
-
-											<__TextAreaContainer>
-												<__InputLabel>내용</__InputLabel>
-
-												<__TextArea
-													{...register('content', {
-														required: '내용을 입력해주세요.',
-														validate: {
-															emptyCheck: (value) =>
-																value.trim().length < 1
-																	? '내용을 입력해주세요.'
-																	: undefined,
-														},
-													})}
-												/>
-											</__TextAreaContainer>
-										</__HeaderInputContainer>
-									</__PostFormInfoCol>
-									<__PostFormControlsCol xs={12} sm={12} md={12} lg={12}>
-										<__PostButtonsContainer>
-											{loading ? (
-												<__Spinner />
-											) : (
-												<>
-													<__SubmitButton
-														type={'button'}
-														onClick={handleSubmit(editPost)}
-														disabled={!isValid}>
-														저장
-													</__SubmitButton>
-													<__CancelButton type={'button'} onClick={clearEditForm}>
-														취소
-													</__CancelButton>
-													<__DeleteButton type={'button'} onClick={handleDeletePost}>
-														삭제
-													</__DeleteButton>
-												</>
-											)}
-										</__PostButtonsContainer>
-									</__PostFormControlsCol>
-								</__PostFormRow>
-							</__Form>
-						</__PostFormContainer>
+						<PostCardEdit post={post} setIsEditing={setIsEditing} />
 					) : (
 						<__PostContainer fluid={'md'}>
-							<__PostImageContainer>
-								<__PostImage src={post.img} />
-							</__PostImageContainer>
+							<__PostImageContainer>{post.img && <__PostImage src={post.img} />}</__PostImageContainer>
 
 							<__HeaderContainer>
 								<__TitleWrapper>
@@ -396,26 +160,6 @@ export default function PostCard({ post }: PostCardProps) {
 					)}
 				</__ModalBody>
 			</__Modal>
-
-			<__ImageCropperModal
-				show={showCropperModal}
-				backdrop={'static'}
-				keyboard={false}
-				size={'xl'}
-				dialogClassName={'custom-modal'}>
-				<__ImageCropperModalBody>
-					{selectedImage && (
-						<ImageCropper
-							imgFile={selectedImage}
-							setSelectedImage={setSelectedImage}
-							setCroppedImage={setCroppedImage}
-							closeImageCropper={closeImageCropper}
-							cropper={cropper}
-							setCropper={setCropper}
-						/>
-					)}
-				</__ImageCropperModalBody>
-			</__ImageCropperModal>
 		</>
 	);
 }
